No longer require N+1 threads.
I used to spin up a single thread just to pull items from the input Iterator
and push them into a channel. But I had to write special MPMC logic for the
channel to be read from multiple threads, which required a shared mutex. So...
why not drop the channel and just use a shared mutex on the input iterator?

Much simpler, and doesn't require the extra thread. If you actually *want* the
caching behavior of pulling items out of the iterator in a separate thread, you
can (eventually) just run that through a single-thread executor w/ an output
cache size to your liking.