Use crossbeam_channel instead of std::sync::mpsc.
I really didn't expect this to be a bottleneck. The input itreator
is a shared Arc<Mutex<Iterator>>, and I expected fighting over the lock
for that was the bottleneck. Boy was I wrong!

With mpsc:
test sending_and_receiving ... bench: 122,697,790 ns/iter (+/- 22,716,060)

With crossbeam_channel:
test sending_and_receiving ... bench:  21,863,243 ns/iter (+/- 1,364,702)

A big thanks to the crossbeam_channel author(s) for bringing back the
ability to know when all senders/receivers have been closed. That was
previously a blocker for me using crossbeam_channel v0.2.